backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
shap>=0.43.0
matplotlib>=3.7.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
streamlit>=1.28.0
lime>=0.2.0
